Move Configuration to fr.ip.model.core;Some cleanup
