+++ src/main/java/fr/ip/model/core/Configuration.java
@@ -1,17 +1,35 @@
-package fr.ip.view.core.components;
+package fr.ip.model.core;
 
 public class Configuration {
 
+    /**
+     * Configuration instance for the current game
+     */
 		public static final Configuration configuration = new Configuration ();
 
+    /**
+     * Global configuration
+     */
     public static class Global {}
 
+    /**
+     * Configuration for either Numeri or Goose game
+     */
     public static class Game {
 
+        /**
+         * Enum the way to display cells in the frame
+         */
 				public static class CellOrder {
 
+            /**
+             * Options' name
+             */
 						public final static String[] options = {"Column", "Rectangle", "Zigzag", "Spiral"};
 
+            /**
+             * Options' symbolic variables
+             */
 						public static final int COLUMN    = 0,
 									 									RECTANGLE = 1,
 																		ZIGZAG    = 2,
@@ -28,15 +46,27 @@ public class Configuration {
         }
     }
 
+    /**
+     * Configuration for goose game
+     */
     public static class Goose extends Game {
 
+        /**
+         * Way to end the game
+         */
         public static class EndMode {
             
+            /**
+             *  Options' name
+             */
             public static final String[] modes = {
                 "Terminer quand sur la dernière case",
                 "Terminer quand dépasse la dernière case"
-            };  
+            };
 
+            /**
+             * Options' symbolic variable
+             */
             public static final int 
               ON_LAST   = 0,
               OVER_LAST = 1;
@@ -44,6 +74,11 @@ public class Configuration {
         }
 
         public int endMode;
+
+        /**
+         * cohabits: Can two or more pawn be on the same cell
+         * questions: ask questions and can replay if correct answer
+         */
         public boolean cohabits, questions;
         
         public Goose (int nC, int nP, int cO, int endMode, boolean cohabits, boolean questions) {
@@ -55,8 +90,15 @@ public class Configuration {
 
     }
 
+    /**
+     * Option for the numeri
+     */
     public static class Numeri extends Game {
         
+        /**
+         * align: replay if 3 pawns are align
+         * replay: additional face on the dice
+         */
         public boolean align, replay;
 
         public Numeri (int nC, int nP, int cO, boolean align, boolean replay) {
@@ -67,6 +109,9 @@ public class Configuration {
 
     }
 
+    /**
+     * Specific configuration instances
+     */
     public final Goose goose;
     public final Numeri numeri;
