+++ src/main/java/fr/ip/model/goose/GoosePlayer.java
@@ -7,7 +7,7 @@ import fr.ip.model.core.Player;
 import fr.ip.model.util.Facade;
 import fr.ip.model.util.Message;
 
-import fr.ip.view.core.components.Configuration;
+import fr.ip.model.core.Configuration;
 
 import javax.swing.*;
